ui: allow main text to be higher by reworking the view and edit
see https://github.com/elabftw/elabftw/discussions/5713
