+++ src/templates/catstat-view.html
@@ -5,19 +5,3 @@
   </div>
   {{ Entity.entityData.team_name }}
 </div>
-
-{# CATEGORY #}
-<div class='d-flex mb-2 align-items-center'>
-  <div class='edit-mode-label'>
-    <i aria-hidden='true' class='fas fa-fw fa-folder mr-1'></i>{{ 'Category'|trans }}
-  </div>
-  <button style='--bg: #{{ Entity.entityData.category_color|default('bdbdbd') }}' type='button' data-target='category' class='btn catstat-btn category-btn malleableCategory lh-normal border-0'>{{ Entity.entityData.category_title|default('Not set'|trans) }}</button>
-</div>
-
-{# STATUS #}
-<div id='statusViewDiv' class='d-flex mb-2 align-items-center'>
-  <div class='edit-mode-label'>
-    <i aria-hidden='true' class='fas fa-fw fa-check-square mr-1'></i>{{ 'Status'|trans }}
-  </div>
-  <button data-target='status' type='button' class='btn catstat-btn status-btn malleableStatus lh-normal border-0'><i class='fas fa-circle fa-fw' style='--bg: #{{ Entity.entityData.status_color|default('bdbdbd') }}'></i> {{ Entity.entityData.status_title|default('Not set'|trans) }}</button>
-</div>
