+++ src/templates/edit-permissions.html
@@ -20,4 +20,3 @@
     {% include('view-permissions.html') %}
   </div>
 {% endif %}
-<hr>
