+++ src/templates/view.html
@@ -4,22 +4,23 @@
 
 {% embed 'view-edit.html' %}
   {% block createnew %}
-    <div class='d-flex'>
-      <div class='align-self-center flex-grow-1 mr-2'>
-        {# CUSTOM ID #}
-        {# TITLE #}
-        <h1 class='text-dark' >
-          {% if Entity.entityData.custom_id %}
-            <span title='{{ 'Custom ID'|trans }}' class='custom-id-badge'>{{ Entity.entityData.custom_id }}</span>
-          {% endif %}
-           <!-- [html-validate-disable-block prefer-native-element: suppress error from malle] -->
-          <span id='documentTitle' class='breakable {{ not Entity.isReadOnly ? 'malleableTitle hl-hover-gray' }} rounded py-1' data-target='title' data-endpoint='{{ Entity.entityType.value }}' data-id='{{ Entity.id }}'>{{ Entity.entityData.title }}</span>
-        </h1>
-      </div>
-      <div class='ml-auto'>
-        {% include 'create-new.html' %}
-      </div>
+  {% include 'catstat-view-edit.html' %}
+  <div class='d-flex'>
+    <div class='align-self-center flex-grow-1 mr-2'>
+      {# CUSTOM ID #}
+      {# TITLE #}
+      <h1 class='text-dark' >
+        {% if Entity.entityData.custom_id %}
+          <span title='{{ 'Custom ID'|trans }}' class='custom-id-badge'>{{ Entity.entityData.custom_id }}</span>
+        {% endif %}
+         <!-- [html-validate-disable-block prefer-native-element: suppress error from malle] -->
+        <span id='documentTitle' class='breakable {{ not Entity.isReadOnly ? 'malleableTitle hl-hover-gray' }} rounded py-1' data-target='title' data-endpoint='{{ Entity.entityType.value }}' data-id='{{ Entity.id }}'>{{ Entity.entityData.title }}</span>
+      </h1>
     </div>
+    <div class='ml-auto'>
+      {% include 'create-new.html' %}
+    </div>
+  </div>
   {% endblock %}
   {% block createmodal %}
     {% include 'show-view-edit.html' %}
@@ -56,7 +57,7 @@
   <div class='d-flex mb-1'>
     {# DATE #}
     {% if Entity.entityType.value != 'experiments_templates' %}
-      <div class='color-medium'>
+      <div class='color-medium mb-2'>
         {{ 'Started on'|trans }} {{ Entity.entityData.date|date('Y-m-d') }}
       </div>
     {% endif %}
@@ -84,16 +85,7 @@
     {% endif %}
   </div>
 
-  {# ID #}
-  <div class='d-flex mb-2 align-items-center'>
-    <div class='edit-mode-label'>
-      <i class='fas fa-hashtag mr-2'></i>{{ 'ID'|trans }}
-    </div>
-    {{ Entity.entityData.id }}
-  </div>
-
   <hr>
-  {% include('edit-permissions.html') %}
 
   {# SHOW EVENTS #}
   {% if Entity.entityData.events_start %}
@@ -125,23 +117,18 @@
       {# body in view mode is text-dark for more lisibility see github issue #52 #}
       <div {{ not displayMainText ? 'hidden' }}>
         <h3 data-action='toggle-next' data-opened-icon='fa-caret-down' data-closed-icon='fa-caret-right' class='d-inline togglable-section-title' tabindex='0' role='button' aria-expanded='true' aria-controls='body_view'><i class='fas fa-caret-down fa-fw mr-2'></i>{{ 'Main text'|trans }}</h3>
-        <div id='body_view' class='pt-2 text-dark' data-save-hidden='body_view'>{{ body|raw }}</div>
+        <div id='body_view' class='pt-2 ml-3 text-dark' data-save-hidden='body_view'>{{ body|raw }}</div>
         <hr>
       </div>
   {% endif %}
 
 </section>
-<div class='d-flex flex-column align-items-end'>
-  {# SHOW OWNER #}
-  <div>{{ 'Owner:'|trans }} {{ Entity.entityData.fullname }}</div>
-  {# SHOW LAST MOD #}
-  <div>{{ 'Last modified on %s'|trans|format(Entity.entityData.modified_at) }}</div>
-  {# eLabID: add an if because templates don't have one #}
-  {% if Entity.entityData.elabid %}
-    <div class='elabid'>{{ 'Unique eLabID:'|trans }} {{ Entity.entityData.elabid }}</div>
-  {% endif %}
-</div>
 
+{# PERMISSIONS #}
+<h3 title='{{ 'Toggle visibility'|trans }}' data-action='toggle-next' data-opened-icon='fa-caret-down' data-closed-icon='fa-caret-right' class='d-inline togglable-section-title' tabindex='0' role='button' aria-expanded='true' aria-controls='permissionsDiv'><i class='fas fa-caret-down fa-fw mr-2'></i>{{ 'Permissions'|trans }}</h3>
+<div class='mt-2 ml-3' id='permissionsDiv' data-save-hidden='permissionsDiv'>
+  {% include('edit-permissions.html') %}
+</div>
 <hr>
 
 {# METADATA view #}
@@ -188,6 +175,19 @@
       </div>
     </div>
 {% endif %}
+{# STATIC INFO #}
+<hr>
+<div class='d-flex flex-column align-items-end'>
+  {# SHOW OWNER #}
+  <div>{{ 'Owner:'|trans }} {{ Entity.entityData.fullname }}</div>
+  {# SHOW LAST MOD #}
+  <div>{{ 'Last modified on %s'|trans|format(Entity.entityData.modified_at) }}</div>
+  {# eLabID: add an if because templates don't have one #}
+  {% if Entity.entityData.elabid %}
+    <div class='elabid'>{{ 'Unique eLabID:'|trans }} {{ Entity.entityData.elabid }}</div>
+  {% endif %}
+</div>
+
 
 <div id='info'
     data-page='view'
